SQUASH ME: follow review hints
