+++ quarkus-extension/simple-rest-example/src/main/java/org/camunda/bpm/quarkus/example/MyServiceDelegate.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.acme;
+package org.camunda.bpm.quarkus.example;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
